requirements.txt
Flask
Gunicorn
openCV-python
Pillow
rembg